+++ src/components/data-management/ProjectCard.jsx
@@ -1,14 +1,29 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import {
   Card, Descriptions,
 } from 'antd';
 import { blue } from '@ant-design/colors';
-import { useSelector } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 import EditableField from '../EditableField';
 import PrettyTime from '../PrettyTime';
 
+import validateInputs, { rules } from '../../utils/validateInputs';
+import ProjectDeleteModal from './ProjectDeleteModal';
+import { setActiveProject, updateProject } from '../../redux/actions/projects';
+
 const { Item } = Descriptions;
+
+const validationChecks = [
+  rules.MIN_8_CHARS,
+  rules.MIN_2_SEQUENTIAL_CHARS,
+  rules.ALPHANUM_DASH_SPACE,
+  rules.UNIQUE_NAME_CASE_INSENSITIVE,
+];
+
+const inactiveProjectStyle = {
+  cursor: 'pointer',
+};
 
 const activeProjectStyle = {
   backgroundColor: blue[0],
@@ -17,82 +32,105 @@ const activeProjectStyle = {
 };
 
 const ProjectCard = (props) => {
-  const {
-    uuid, isActive, onClick, onSubmit, onDelete, validationFn,
-  } = props;
+  const { projectUuid } = props;
 
-  const project = useSelector((state) => state.projects[uuid]);
+  const dispatch = useDispatch();
+
+  const projects = useSelector((state) => state.projects);
+
+  const [deleteModalVisible, setDeleteModalVisible] = useState(false);
+
+  const { activeProjectUuid } = projects.meta;
+  const project = projects[projectUuid];
+
+  const projectNames = projects.ids.map((uuid) => projects[uuid].name);
+
+  const validationParams = {
+    existingNames: projectNames,
+  };
+
+  const updateProjectName = (newName) => {
+    dispatch(updateProject(project.uuid, { name: newName }));
+  };
+
+  const deleteProject = () => {
+    setDeleteModalVisible(true);
+  };
 
   return (
-    <Card
-      data-test-class='project-card'
-      key={uuid}
-      type='primary'
-      style={isActive ? activeProjectStyle : { cursor: 'pointer' }}
-      onClick={onClick}
-    >
-      <Descriptions
-        layout='horizontal'
-        size='small'
-        column={1}
+    <>
+      {deleteModalVisible && (
+        <ProjectDeleteModal
+          projectUuid={project.uuid}
+          onCancel={() => { setDeleteModalVisible(false); }}
+          onDelete={() => { setDeleteModalVisible(false); }}
+        />
+      )}
+      <Card
+        data-test-class='project-card'
+        key={projectUuid}
+        type='primary'
+        style={projectUuid === activeProjectUuid ? activeProjectStyle : inactiveProjectStyle}
+        onClick={() => {
+          dispatch(setActiveProject(project.uuid));
+        }}
       >
-        <Item contentStyle={{ fontWeight: 700, fontSize: 16 }}>
-          <EditableField
-            value={project.name}
-            onAfterSubmit={onSubmit}
-            onDelete={onDelete}
-            validationFunc={validationFn}
-          />
-        </Item>
-        <Item
-          labelStyle={{ fontWeight: 'bold' }}
-          label='Samples'
+        <Descriptions
+          layout='horizontal'
+          size='small'
+          column={1}
         >
-          {project.samples.length}
+          <Item contentStyle={{ fontWeight: 700, fontSize: 16 }}>
+            <EditableField
+              value={project.name}
+              onAfterSubmit={updateProjectName}
+              onDelete={deleteProject}
+              validationFunc={
+                (newName) => validateInputs(
+                  newName,
+                  validationChecks,
+                  validationParams,
+                ).isValid
+              }
+            />
+          </Item>
+          <Item
+            labelStyle={{ fontWeight: 'bold' }}
+            label='Samples'
+          >
+            {project.samples.length}
 
-        </Item>
-        <Item
-          labelStyle={{ fontWeight: 'bold' }}
-          label='Created'
-        >
-          <PrettyTime isoTime={project.createdDate} />
+          </Item>
+          <Item
+            labelStyle={{ fontWeight: 'bold' }}
+            label='Created'
+          >
+            <PrettyTime isoTime={project.createdDate} />
 
-        </Item>
-        <Item
-          labelStyle={{ fontWeight: 'bold' }}
-          label='Modified'
-        >
-          <PrettyTime isoTime={project.lastModified} />
+          </Item>
+          <Item
+            labelStyle={{ fontWeight: 'bold' }}
+            label='Modified'
+          >
+            <PrettyTime isoTime={project.lastModified} />
 
-        </Item>
-        <Item
-          labelStyle={{ fontWeight: 'bold' }}
-          label='Last analyzed'
-        >
-          {project.lastAnalyzed ? (
-            <PrettyTime isoTime={project.lastAnalyzed} />
-          ) : ('never')}
-        </Item>
-      </Descriptions>
-    </Card>
+          </Item>
+          <Item
+            labelStyle={{ fontWeight: 'bold' }}
+            label='Last analyzed'
+          >
+            {project.lastAnalyzed ? (
+              <PrettyTime isoTime={project.lastAnalyzed} />
+            ) : ('never')}
+          </Item>
+        </Descriptions>
+      </Card>
+    </>
   );
 };
 
 ProjectCard.propTypes = {
-  uuid: PropTypes.string.isRequired,
-  isActive: PropTypes.bool,
-  onClick: PropTypes.func,
-  onSubmit: PropTypes.func,
-  onDelete: PropTypes.func,
-  validationFn: PropTypes.func,
-};
-
-ProjectCard.defaultProps = {
-  isActive: false,
-  onClick: () => {},
-  onSubmit: () => {},
-  onDelete: () => {},
-  validationFn: () => {},
+  projectUuid: PropTypes.string.isRequired,
 };
 
 export default ProjectCard;
