+++ src/components/data-management/ProjectDeleteModal.jsx
@@ -1,16 +1,21 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
+import { useDispatch, useSelector } from 'react-redux';
 import {
   Modal, Button, Input, Space, Typography, Form, Alert,
 } from 'antd';
 import integrationTestConstants from '../../utils/integrationTestConstants';
+import deleteProject from '../../redux/actions/projects/deleteProject';
 
 const { Text, Paragraph } = Typography;
 
 const ProjectDeleteModal = (props) => {
   const {
-    visible, onDelete, onCancel, projectName,
+    projectUuid, onCancel, onDelete,
   } = props;
+
+  const dispatch = useDispatch();
+  const projectName = useSelector((state) => state.projects[projectUuid].name);
 
   const [inputProjectName, setInputProjectName] = useState('');
   const [isValid, setIsValid] = useState(false);
@@ -19,15 +24,14 @@ const ProjectDeleteModal = (props) => {
     <Modal
       className={integrationTestConstants.classes.DELETE_PROJECT_MODAL}
       title='Confirm delete'
-      visible={visible}
+      visible
       footer={(
         <Space>
           <Button
             type='secondary'
             key='cancel'
             onClick={() => {
-              onCancel(inputProjectName);
-              setInputProjectName('');
+              onCancel();
               setIsValid(false);
             }}
           >
@@ -39,9 +43,8 @@ const ProjectDeleteModal = (props) => {
             key='create'
             disabled={!isValid}
             onClick={() => {
+              dispatch(deleteProject(projectUuid));
               onDelete();
-              setInputProjectName('');
-              setIsValid(false);
             }}
           >
             Permanently delete project
@@ -101,17 +104,14 @@ const ProjectDeleteModal = (props) => {
 };
 
 ProjectDeleteModal.propTypes = {
-  visible: PropTypes.bool,
-  onDelete: PropTypes.func,
+  projectUuid: PropTypes.string.isRequired,
   onCancel: PropTypes.func,
-  projectName: PropTypes.string,
+  onDelete: PropTypes.func,
 };
 
 ProjectDeleteModal.defaultProps = {
-  onDelete: () => null,
   onCancel: () => null,
-  visible: false,
-  projectName: null,
+  onDelete: () => null,
 };
 
 export default ProjectDeleteModal;
